requirements.txt
discord.py
pandas==1.0.5